==> notebooks/02_Create_Predictions.ipynb <==
# 02 Create Predictions
Create predictions for each potential game

import pandas as pd
import numpy as np

## Read in data

comb_results = pd.read_csv('../data/comb_results.csv')
comb_results.head()

seeds = pd.read_csv('../data/Stage2DataFiles//NCAATourneySeeds.csv')
seeds['Seed_v2'] = [int(x[1:3]) for x in seeds['Seed']]
seeds.drop('Seed', axis=1, inplace=True)
seeds.head()

## Get performance by seed matchup

# join to get original seed for strong and weak teams
comb_results_seeds = pd.merge(comb_results, seeds, 
                              left_on=['Season', 'TeamID_Strong'], 
                              right_on=['Season', 'TeamID'], 
                              validate='m:1')
comb_results_seeds.rename(columns={'Seed_v2':'Orig_Seed_Strong'}, inplace=True)
comb_results_seeds.drop('TeamID', axis=1, inplace=True)

comb_results_seeds = pd.merge(comb_results_seeds, seeds, 
                              left_on=['Season', 'TeamID_Weak'], 
                              right_on=['Season', 'TeamID'], 
                              validate='m:1')
comb_results_seeds.rename(columns={'Seed_v2':'Orig_Seed_Weak'}, inplace=True)
comb_results_seeds.drop('TeamID', axis=1, inplace=True)

comb_results_seeds.head()

# add win indicator for strong team
comb_results_seeds_v2 = comb_results_seeds.copy()
comb_results_seeds_v2['Strong_Win'] = [1] * (comb_results_seeds_v2['TeamID_Winner']
                                             ==comb_results_seeds_v2['TeamID_Strong'])
comb_results_seeds_v2.head()

comb_results_seeds_v2['comp'] = 1
comb_results_seeds_v2['comp'] = [1 if s<w else -1 for w,s in zip(comb_results_seeds_v2['Orig_Seed_Weak'], 
                                            comb_results_seeds_v2['Orig_Seed_Strong'])]


comb_results_seeds_v2['Strong_Win'] = [p if c==1 else (1-p) for c,p in zip(comb_results_seeds_v2['comp'], comb_results_seeds_v2['Strong_Win'])]


# Aggregate by round and matchup
results_agg = comb_results_seeds_v2.groupby(['Round', 'Orig_Seed_Strong', 'Orig_Seed_Weak']).agg(
    {'Strong_Win':sum, 'Season':'count'}).reset_index()
results_agg.rename(columns={'Season':'Count'}, inplace=True)
results_agg.head()

results_agg.to_csv('../data/results_agg.csv', index=False)

# @np.vectorize
# group(rnd, s, w):
#     x = 1
#     if rnd==1: 
#         x+=1
#         return (x-1)
#     elif rnd==2:
        
        
@np.vectorize
def agg_matchup(w, s):
    min_s = min(w,s)
    max_s = max(w,s)
    diff = max_s - min_s
    return diff

results_agg['Matchup'] = agg_matchup(results_agg['Orig_Seed_Strong'], results_agg['Orig_Seed_Weak'])
results_agg.head()
# Think of ways to get more data points for later rounds.  Maybe group rounds together and seeds together
# and think of way to get combinations instead of permutations

results_agg_v2 = results_agg.groupby(['Round', 'Matchup']).agg({'Strong_Win':sum, 'Count':sum}).reset_index()
results_agg_v2.head()

@np.vectorize
def agg_round(r):
    if r<3:
        return r
    elif r>=3:
        return 3
    
results_agg_v2['Round_v2'] = agg_round(results_agg_v2['Round'])
results_agg_v2.head()

results_agg_v3 = results_agg_v2.groupby(['Round_v2', 'Matchup']).agg({'Strong_Win':sum, 'Count':sum}).reset_index()
results_agg_v3.head()

results_agg_v3['Key'] = [r*100 + m for r,m in zip(results_agg_v3['Round_v2'], results_agg_v3['Matchup'])]
results_agg_v4 = results_agg_v3.groupby('Key').agg({'Strong_Win':sum, 'Count':sum}).reset_index()
results_agg_v4['Win_Prob'] = results_agg_v4['Strong_Win']/results_agg_v4['Count']
results_agg_v4.head()

results_agg_v4['Win_Prob'] = [0.5 if c<10 else w for c,w in zip(results_agg_v4['Count'], results_agg_v4['Win_Prob'])]

results_agg_v4.to_csv('../data/model_v1.csv', index=False)

results_agg_v4
