==> notebooks/02_Create_Predictions_GBM.ipynb <==
# 02 Create Predictions GBM
Create predictions for each potential game

import pandas as pd
import numpy as np
from xgboost.sklearn import XGBClassifier
import pickle

## Read in data

comb_results = pd.read_csv('../data/comb_results.csv')
comb_results.head()

seeds = pd.read_csv('../data/Stage2DataFiles//NCAATourneySeeds.csv')
seeds['Seed_v2'] = [int(x[1:3]) for x in seeds['Seed']]
seeds.drop('Seed', axis=1, inplace=True)
seeds.head()

ordinals = pd.read_csv('../data/MasseyOrdinals_thru_2019_day_128/MasseyOrdinals_thru_2019_day_128.csv')
rpi = ordinals.loc[ordinals['SystemName']=='RPI']
rpi = rpi.drop_duplicates(subset=['Season', 'SystemName', 'TeamID'], keep='last')
rpi.head()

## Create features

# join to get original seed for strong and weak teams
comb_results_seeds = pd.merge(comb_results, seeds, 
                              left_on=['Season', 'TeamID_Strong'], 
                              right_on=['Season', 'TeamID'], 
                              validate='m:1')
comb_results_seeds.rename(columns={'Seed_v2':'Orig_Seed_Strong'}, inplace=True)
comb_results_seeds.drop('TeamID', axis=1, inplace=True)

comb_results_seeds = pd.merge(comb_results_seeds, seeds, 
                              left_on=['Season', 'TeamID_Weak'], 
                              right_on=['Season', 'TeamID'], 
                              validate='m:1')
comb_results_seeds.rename(columns={'Seed_v2':'Orig_Seed_Weak'}, inplace=True)
comb_results_seeds.drop('TeamID', axis=1, inplace=True)

comb_results_seeds.head()

# add win indicator for strong team
comb_results_seeds_v2 = comb_results_seeds.copy()
comb_results_seeds_v2['Strong_Win'] = [1] * (comb_results_seeds_v2['TeamID_Winner']
                                             ==comb_results_seeds_v2['TeamID_Strong'])
comb_results_seeds_v2.head()

# add ordinals
comb_results_seeds_v3 = pd.merge(comb_results_seeds_v2, rpi[['Season', 'TeamID', 'OrdinalRank']], 
                           left_on=['Season', 'TeamID_Strong'], 
                           right_on=['Season', 'TeamID'],
                           how='left', validate='m:1')
comb_results_seeds_v3.rename(columns={'OrdinalRank':'RPI_Strong'}, inplace=True)
comb_results_seeds_v3.drop(['TeamID'], axis=1, inplace=True)

comb_results_seeds_v3 = pd.merge(comb_results_seeds_v3, rpi[['Season', 'TeamID', 'OrdinalRank']], 
                           left_on=['Season', 'TeamID_Weak'], 
                           right_on=['Season', 'TeamID'], 
                           how='left', validate='m:1')
comb_results_seeds_v3.rename(columns={'OrdinalRank':'RPI_Weak'}, inplace=True)
comb_results_seeds_v3.drop(['TeamID'], axis=1, inplace=True)

# create dummy variable
comb_results_seeds_v3['RPI_Strong_Dummy'] = [0 if pd.isnull(x) else 1 for x in comb_results_seeds_v3['RPI_Strong']]
comb_results_seeds_v3['RPI_Weak_Dummy'] = [0 if pd.isnull(x) else 1 for x in comb_results_seeds_v3['RPI_Weak']]

comb_results_seeds_v3['RPI_Weak'] = comb_results_seeds_v3['RPI_Weak'].fillna(0)
comb_results_seeds_v3['RPI_Strong'] = comb_results_seeds_v3['RPI_Strong'].fillna(0)

## Build Model

y = comb_results_seeds_v3[['Strong_Win']]
X = comb_results_seeds_v3[['Orig_Seed_Strong', 'Orig_Seed_Weak', 'RPI_Strong', 'RPI_Weak', 
                           'RPI_Strong_Dummy', 'RPI_Weak_Dummy']]

Xt = pd.DataFrame()
Xt['Orig_Seed_Weak'] = X['Orig_Seed_Strong']
Xt['Orig_Seed_Strong'] = X['Orig_Seed_Weak']
Xt['RPI_Strong'] = X['RPI_Weak']
Xt['RPI_Strong_Dummy'] = X['RPI_Weak_Dummy']
Xt['RPI_Weak'] = X['RPI_Strong']
Xt['RPI_Weak_Dummy'] = X['RPI_Strong_Dummy']

Xt.head()

yt = [1-y]

X_all = X.append(Xt).reset_index(drop=True)
y_all = y.append(yt).reset_index(drop=True)

reg = XGBClassifier().fit(X_all, y_all)

pred = X.copy()[['Orig_Seed_Strong', 'Orig_Seed_Weak', 'RPI_Strong', 'RPI_Strong_Dummy', 'RPI_Weak', 'RPI_Weak_Dummy']]
pred['pred'] = [x[1] for x in reg.predict_proba(pred)]
pred.head()

### Save model to file

with open('../data/xgb.p', 'wb') as handle:
    pickle.dump(reg, handle)
