==> notebooks/03_Create_Opt_Input.ipynb <==
import pandas as pd
import numpy as np
import pickle
import sys
sys.path.append('../src')
from data_prep import simulate_round_results, create_round_results

## Import data

results_compact = pd.read_csv('../data/Stage2DataFiles/NCAATourneyCompactResults.csv')
results_compact.head()

teams = pd.read_csv('../data/Stage2DataFiles/teams.csv')
teams.head()

tourney_slots = pd.read_csv('../data/Stage2DataFiles/NCAATourneySlots.csv')
tourney_slots.head()

tourney_seeds = pd.read_csv('../data/Stage2DataFiles/NCAATourneySeeds.csv')
tourney_seeds.head()

seeds = pd.read_csv('../data/Stage2DataFiles//NCAATourneySeeds.csv')
seeds['Seed_v2'] = [int(x[1:3]) for x in seeds['Seed']]
seeds.drop('Seed', axis=1, inplace=True)
seeds.head()

tourney_seed_round = pd.read_csv('../data/Stage2DataFiles/NCAATourneySeedRoundSlots.csv')
tourney_seed_round.head()

ordinals = pd.read_csv('../data/MasseyOrdinals_thru_2019_day_128/MasseyOrdinals_thru_2019_day_128.csv')
rpi = ordinals.loc[ordinals['SystemName']=='RPI']
rpi = rpi.drop_duplicates(subset=['Season', 'SystemName', 'TeamID'], keep='last')
rpi.head()

### Load models

with open('../data/xgb.p', 'rb') as handle:
    xgb_model = pickle.load(handle)

## Data Prep

# Eliminate 2019 season and add round variable
tourney_slots_v2 = tourney_slots.copy()

print (len(tourney_slots_v2))
tourney_slots_v2 = tourney_slots_v2.loc[tourney_slots_v2['Season']!=2019]
print (len(tourney_slots_v2))

@np.vectorize
def add_round(slot):
    if slot[0:1]!='R':
        return 0
    else:
        return int(slot.split('R')[1][0])

tourney_slots_v2['Round'] = add_round(tourney_slots_v2['Slot'])
display(tourney_slots_v2['Round'].value_counts())
tourney_slots_v2.head()

# convert results to season, round, team format
tourney_seed_early_day = tourney_seed_round.groupby('GameRound').agg({'EarlyDayNum':'mean'}).reset_index()
tourney_seed_late_day = tourney_seed_round.groupby('GameRound').agg({'LateDayNum':'mean'}).reset_index()


results_round = pd.merge(results_compact, tourney_seed_early_day, 
                         left_on=['DayNum'], right_on=['EarlyDayNum'], 
                         how='left', 
                         validate='m:1')

results_round = pd.merge(results_round, tourney_seed_late_day, 
                         left_on=['DayNum'], right_on=['LateDayNum'], 
                         how='left', 
                         validate='m:1')

assert len(results_round.loc[
    (pd.isnull(results_round['GameRound_x'])) & 
    (pd.isnull(results_round['GameRound_y']))]
          )==0

valids = results_round.loc[
    (~pd.isnull(results_round['GameRound_x'])) & 
    (~pd.isnull(results_round['GameRound_y']))]
assert len(valids.loc[valids['GameRound_x']!=valids['GameRound_y']])==0

results_round['Round'] = results_round['GameRound_x'].combine_first(results_round['GameRound_y'])

results_w = results_round[['Season', 'Round', 'WTeamID', 'WScore', 'NumOT']].rename(columns={'WTeamID':'TeamID', 'WScore':'Score'})
results_l = results_round[['Season', 'Round', 'LTeamID', 'LScore', 'NumOT']].rename(columns={'LTeamID':'TeamID', 'LScore':'Score'})

results = results_w.append(results_l).reset_index(drop=True)
results.head()

## Simulate Rounds

r0_results = create_round_results(tourney_seeds, 0, teams, tourney_slots_v2, results)
r0_results.head()

r0_results = r0_results.loc[r0_results['Season']==2018]
r0_results

r1_seeds = r0_results[['Season', 'TeamID_Winner', 'Slot']].copy()
r1_seeds.rename(columns={'TeamID_Winner':'TeamID', 'Slot':'Seed'}, inplace=True)

play_ins = list(r1_seeds['Seed'].values)
# play_ins

tourney_seeds_no_play_in = tourney_seeds.copy()
for s in play_ins:
    tourney_seeds_no_play_in = tourney_seeds_no_play_in.loc[tourney_seeds_no_play_in['Seed']!=s+'a']
    tourney_seeds_no_play_in = tourney_seeds_no_play_in.loc[tourney_seeds_no_play_in['Seed']!=s+'b']

r1_seeds = r1_seeds.append(tourney_seeds_no_play_in).reset_index(drop=True)

r1_seeds = r1_seeds.loc[r1_seeds['Season']==2018]

print (r1_seeds.shape)
r1_seeds.head()

r1_results = simulate_round_results(r1_seeds, 1, teams, tourney_slots_v2)
r1_results.head()

r2_seeds = r1_results[['Season', 'TeamID_Winner', 'Slot']].copy()
r2_seeds.rename(columns={'TeamID_Winner':'TeamID', 'Slot':'Seed'}, inplace=True)
r2_seeds.head()

r2_seeds.shape

r2_results = simulate_round_results(r2_seeds, 2, teams, tourney_slots_v2)
r2_results.head()

r2_results.shape

r3_seeds = r2_results[['Season', 'TeamID_Winner', 'Slot']].copy()
r3_seeds.rename(columns={'TeamID_Winner':'TeamID', 'Slot':'Seed'}, inplace=True)
r3_seeds.head()

r3_seeds.shape
r3_seeds.drop_duplicates(inplace=True)
r3_seeds.shape

r3_results = simulate_round_results(r3_seeds, 3, teams, tourney_slots_v2)
r3_results.head()

r3_results.shape

r4_seeds = r3_results[['Season', 'TeamID_Winner', 'Slot']].copy()
r4_seeds.rename(columns={'TeamID_Winner':'TeamID', 'Slot':'Seed'}, inplace=True)
r4_seeds.head()

r4_seeds.shape
r4_seeds.drop_duplicates(inplace=True)
r4_seeds.shape

r4_results = simulate_round_results(r4_seeds, 4, teams, tourney_slots_v2)
r4_results.head()

r4_results.shape

r5_seeds = r4_results[['Season', 'TeamID_Winner', 'Slot']].copy()
r5_seeds.rename(columns={'TeamID_Winner':'TeamID', 'Slot':'Seed'}, inplace=True)
r5_seeds.head()

r5_seeds.shape
r5_seeds.drop_duplicates(inplace=True)
r5_seeds.shape

r5_results = simulate_round_results(r5_seeds, 5, teams, tourney_slots_v2)
r5_results.head()

r6_seeds = r5_results[['Season', 'TeamID_Winner', 'Slot']].copy()
r6_seeds.rename(columns={'TeamID_Winner':'TeamID', 'Slot':'Seed'}, inplace=True)
r6_seeds.head()

r6_seeds.shape
r6_seeds.drop_duplicates(inplace=True)
r6_seeds.shape

r6_results = simulate_round_results(r6_seeds, 6, teams, tourney_slots_v2)
r6_results.head()

## Combine results

comb_results = r1_results
for d in [r2_results, r3_results, r4_results, r5_results, r6_results]:
    comb_results = comb_results.append(d)
comb_results = comb_results.reset_index(drop=True)
print (comb_results.shape)
comb_results.head()

## Join with original seed

# join to get original seed for strong and weak teams
comb_results_seeds = pd.merge(comb_results, seeds, 
                              left_on=['Season', 'TeamID_Strong'], 
                              right_on=['Season', 'TeamID'], 
                              validate='m:1')
comb_results_seeds.rename(columns={'Seed_v2':'Orig_Seed_Strong'}, inplace=True)
comb_results_seeds.drop('TeamID', axis=1, inplace=True)

comb_results_seeds = pd.merge(comb_results_seeds, seeds, 
                              left_on=['Season', 'TeamID_Weak'], 
                              right_on=['Season', 'TeamID'], 
                              validate='m:1')
comb_results_seeds.rename(columns={'Seed_v2':'Orig_Seed_Weak'}, inplace=True)
comb_results_seeds.drop('TeamID', axis=1, inplace=True)

comb_results_seeds.head()

## Get key info

@np.vectorize
def agg_matchup(w, s):
    min_s = min(w,s)
    max_s = max(w,s)
    diff = max_s - min_s
    return diff

comb_results_seeds['Matchup'] = agg_matchup(comb_results_seeds['Orig_Seed_Weak'], 
                                            comb_results_seeds['Orig_Seed_Strong'])

@np.vectorize
def agg_round(r):
    if r<3:
        return r
    elif r>=3:
        return 3
    
comb_results_seeds['Round_v2'] = agg_round(comb_results_seeds['Round'])
comb_results_seeds.head()

comb_results_seeds['comp'] = 1
comb_results_seeds['comp'] = [1 if s<w else -1 for w,s in zip(comb_results_seeds['Orig_Seed_Weak'], 
                                            comb_results_seeds['Orig_Seed_Strong'])]
comb_results_seeds['comp'] = comb_results_seeds['comp']*[1 if w==1 else -1 for w in comb_results_seeds['Strong_Win']]


comb_results_seeds['Key'] = [r*100 + m for r,m in zip(comb_results_seeds['Round_v2'], 
                                                       comb_results_seeds['Matchup'])]



model_v1 = pd.read_csv('../data/model_v1.csv')
model_v1 = model_v1[['Key', 'Win_Prob']]
model_v1.head()

pred = pd.merge(comb_results_seeds, model_v1, on='Key', how='left', validate='m:1')
pred['Win_Prob'] = [p if c==1 else (1-p) for c,p in zip(pred['comp'], pred['Win_Prob'])]

pred['Win_Prob'] = pred['Win_Prob'].fillna(0.5)

## Add ordinal information

# add ordinals
pred = pd.merge(pred, rpi[['Season', 'TeamID', 'OrdinalRank']], 
                           left_on=['Season', 'TeamID_Strong'], 
                           right_on=['Season', 'TeamID'],
                           how='left', validate='m:1')
pred.rename(columns={'OrdinalRank':'RPI_Strong'}, inplace=True)
pred.drop(['TeamID'], axis=1, inplace=True)

pred = pd.merge(pred, rpi[['Season', 'TeamID', 'OrdinalRank']], 
                           left_on=['Season', 'TeamID_Weak'], 
                           right_on=['Season', 'TeamID'], 
                           how='left', validate='m:1')
pred.rename(columns={'OrdinalRank':'RPI_Weak'}, inplace=True)
pred.drop(['TeamID'], axis=1, inplace=True)

# create dummy variable
pred['RPI_Strong_Dummy'] = [0 if pd.isnull(x) else 1 for x in pred['RPI_Strong']]
pred['RPI_Weak_Dummy'] = [0 if pd.isnull(x) else 1 for x in pred['RPI_Weak']]

pred['RPI_Weak'] = pred['RPI_Weak'].fillna(0)
pred['RPI_Strong'] = pred['RPI_Strong'].fillna(0)

## Add GLM

# @np.vectorize
# def apply_glm(oss, osw):
#     b0 = -7.73792442e-15
#     b1 = -0.16661885
#     b2 = 0.16661885
#     return (1/(1+np.e**(-1*(b0+b1*oss+b2*osw))))

@np.vectorize
def apply_glm(oss, osw, rs, rsd, rw, rwd):
    b0 = 7.78842465e-15
    B = [-1.54628007e-01,  1.54628007e-01, -4.78439446e-03,
         9.31424100e-15,  4.78439446e-03,  9.31424100e-15]
    return (1/(1+np.e**(-1*(b0+B[0]*oss + B[1]*osw + B[2]*rs + B[3]*rsd + B[4]*rw + B[5]*rwd))))

pred['Win_Prob_glm_temp'] = apply_glm(pred['Orig_Seed_Strong'], pred['Orig_Seed_Weak'], pred['RPI_Strong'], pred['RPI_Strong_Dummy'], pred['RPI_Weak'], pred['RPI_Weak_Dummy'])
pred['Win_Prob_glm'] = [p if c==1 else (1-p) for c,p in zip(pred['comp'], pred['Win_Prob_glm_temp'])]

pred.head()

### Add GBM

X_gbm = pred[['Orig_Seed_Strong', 'Orig_Seed_Weak', 'RPI_Strong', 'RPI_Strong_Dummy', 'RPI_Weak', 'RPI_Weak_Dummy']]
pred['Win_Prob_gbm_temp'] = [x[1] for x in xgb_model.predict_proba(X_gbm)]
pred['Win_Prob_gbm'] = [p if c==1 else (1-p) for c,p in zip(pred['comp'], pred['Win_Prob_gbm_temp'])]



pred.head()

pred.to_csv('../data/opt_input.csv', index=False)

pred.loc[(pred['Win_Prob_glm']-pred['Win_Prob_gbm'])>.3]
