==> notebooks/04_Optimize.ipynb <==
# 04 Optimize

import pandas as pd
import numpy as np
import sys
sys.path.append('../src')
from optimize import optimize_tourney

df = pd.read_csv('../data/opt_input.csv')
df.head()

# weight win prob by round
df['Win_Prob_weighted'] = [w*(2**(r-1)) for w,r in zip(df['Win_Prob'], df['Round'])]
df['Win_Prob_glm_weighted'] = [w*(2**(r-1)) for w,r in zip(df['Win_Prob_glm'], df['Round'])]
df['Win_Prob_gbm_weighted'] = [w*(2**(r-1)) for w,r in zip(df['Win_Prob_gbm'], df['Round'])]

df.head()

opt_results = optimize_tourney(df, 'Win_Prob_gbm', 'max')

opt_results.head()

opt_results.to_csv('../data/opt_results.csv', index=False)

df.to_csv('../data/opt_input_v2.csv', index=False)

### Simulate by picking team with highest prob at each round

win_var = 'Win_Prob_glm'
winners = list(df['TeamID_Strong']) + list(df['TeamID_Weak'])
output = pd.DataFrame()

for r in [1,2,3,4,5,6]:
    # filter to round
    rnd = df.loc[df['Round']==r]
    # drop duplicates
    rnd = rnd.drop_duplicates(['Season', 'Slot', 'StrongSeed', 'WeakSeed', 'Round', 'TeamID_Strong', 'TeamID_Weak'])
    # limit down to potential teams
    rnd = rnd.loc[rnd['TeamID_Strong'].isin(winners)]
    rnd = rnd.loc[rnd['TeamID_Weak'].isin(winners)]
    # pick winner
    rnd['TeamID_Winner'] = [s if p>=0.5 else w for s,w,p in zip(rnd['TeamID_Strong'], rnd['TeamID_Weak'], rnd[win_var])]
    winners = list(rnd['TeamID_Winner'].values)
    output = output.append(rnd)

output.to_csv('../data/opt_results_simple.csv', index=False)
