==> notebooks/05_Score_Accuracy.ipynb <==
# 05 Score Accuracy

import pandas as pd
import numpy as np
import sys
sys.path.append('../src')
# from optimize import optimize_tourney

## Read in data

# opt_results = pd.read_csv('../data/opt_results.csv')
opt_results = pd.read_csv('../data/opt_results_simple.csv')

opt_results.head()

actuals = pd.read_csv('../data/comb_results.csv')
actuals.head()

## Filter down data

opt_results_v2 = opt_results[['Season' ,'Slot', 'TeamID_Winner']]
opt_results_v2.head()

actuals_v2 = actuals[['Season', 'Slot', 'Round', 'TeamID_Winner']]
actuals_v2.head()

## Join data

pred_actuals = pd.merge(opt_results_v2, actuals_v2, on=['Season', 'Slot'], suffixes=['_pred', '_actual'], validate='1:1')

pred_actuals['Correct'] = 1*(pred_actuals['TeamID_Winner_pred']==pred_actuals['TeamID_Winner_actual'])
pred_actuals['Points'] = [2**(r-1)*c for r,c in zip(pred_actuals['Round'], pred_actuals['Correct'])]

pred_actuals.head()

print (len(pred_actuals))
print (pred_actuals['Correct'].sum())
print (pred_actuals['Correct'].sum()/len(pred_actuals))
print (pred_actuals['Points'].sum())

# empirical
# 63
# 24
# 0.38095238095238093
# 41

# GLM
# 63
# 27
# 0.42857142857142855
# 61

# GLM with ordinal rankings
# 63
# 23
# 0.36507936507936506
# 55

# GLM with ordinal rankings with simple picking
# 63
# 35
# 0.5555555555555556
# 65

# GBM with ordinal rankings
# 63
# 26
# 0.4126984126984127
# 37

# GBM with simple picking
# 63
# 35
# 0.5555555555555556
# 77

pred_actuals.to_csv('../data/scores.csv', index=False)
